Fix assertion failure when file ends with comment token
With an unterminated comment and no terminating newline, we would
try to read past end of file as lexer_ctext did not check for end
of file. Also the assertion in lexer_advance() was not correct
since we can be at the end of lexer buffer at the end of file.
Found with zzuf fuzzer.