Add parser for the IR language
Now syc will accept an .ir file (in addition to a .c or .h file).
Instruction selection asserts certain properties of the IR that
may not hold if a hand-written IR file is parsed. Since an assertion
failure is not a good failure mode, it would be nice to add some extra
validation step. This must be performed while parsing, because we chose
to forget the source code immediately after parsing and thus we could
not provide references to the IR source in the error messages otherwise.