Warn unused / fix versioned variables
Generate warning for unused local variables and function arguments.
To be able to print source coordinates of the definitions, we need
to store the entire lexer token in the scope instead of just
the identifier.

Versioned variables (i.e. in case of multiple variables of the same
name in the same function) look like %version@name (because of z80asm
constraints), this was being confused with %n virtual registers,
because it starts with a digit after the '%'.