Argument location allocator
A class that determines how each argument should be passed. It uses
a different strategy for 8-bit and for 16-bit (or larger) arguments.
It allows to pass larger (e.g. 32-bit and 64-bit) arguments partially
in registers and partially on the stack.

It would be also possible (not implemented, though) to pass different
parts of an argument in different sized registers (e.g. the elements
of a structure containing 8-bit and 16-bit members).