Add test and fix switch/case
Case condition used the wrong/inverted jump condition. We were not
implementing fall through correctly. This is because we keep the
order of instructions consistent with the source code, i.e., case
condition evaluation is interspersed between the other code. Thus
if we reach a case condition from above, we must jump around it into
the case body.