Function declarators need a separate scope
For this to work, cur_scope needs to be moved from cgen_proc_t to
cgen_t.

This is needed so that structures/unions defined inside parameter list
do not leak out into the parent scope.

It also removes the need to pass extra scope argument with codegen
functions that onlytake cgen and not cgen_proc as the first argument.