Check switch expression type, case value in range
Switch expression can be an integer, truth value or enum. For integers,
test if case value is in range of the type, for logic type check that
case value is 0/1, for enum check that case value is one of the enum
values. The last one may be superfluous if we do strict enum type
checking.

TODO: Check type of case expression value. This is an integer constant
expression, so it can be an integer, truth value or an enum.