Switch/case type compatibility checking
This is specifically because of strict enum and strict truth types.
Commonly, in C all switch and case expressions are effectively integer
types (even if they are bool, enum) and no type checking is done.

We could possibly achieve similar result just by implicitly casting
the value of the case expression to the type of the switch expression.
However, we provide special-case handling here for the benefit of
clearer messaging.